+++ src/webadmin/handlers/mod.rs
@@ -16,6 +16,7 @@ pub mod mfa;
 pub mod misc;
 pub mod orders;
 pub mod paging;
+pub mod params;
 pub mod session;
 
 pub use accounts::*;